// src/routes/step-2/datapribadi.schema.ts
import { z } from 'zod'

export const datapribadiSchema = z.object({
  student_name: z.string({ message: 'nama tidak boleh kosong' }).min(3, { message: 'nama minimal tiga huruf' }),
  student_nik: z.string({ message: 'nama tidak boleh kosong' }),
  student_bop: z.string({ message: 'nama tidak boleh kosong' }),
  student_order_family: z.string({ message: 'nama tidak boleh kosong' }),
  student_gender: z.string({ message: 'nama tidak boleh kosong' }),
  student_height: z.string({ message: 'nama tidak boleh kosong' }),
  student_weight: z.string({ message: 'nama tidak boleh kosong' }),
  student_dress_size: z.string({ message: 'nama tidak boleh kosong' }),
  levelgroup_id: z.string({ message: 'nama tidak boleh kosong' }),
  level_id: z.string({ message: 'nama tidak boleh kosong' }),
  student_nokk: z.string({ message: 'nama tidak boleh kosong' }),
  student_father_name: z.string({ message: 'nama tidak boleh kosong' }),
  student_father_nik: z.string({ message: 'nama tidak boleh kosong' }),
  student_father_status: z.string({ message: 'nama tidak boleh kosong' }),
  //student_father_dob: z.string({ message: 'nama tidak boleh kosong' }),
  student_father_bop: z.string({ message: 'nama tidak boleh kosong' }),
  student_father_lastedu: z.string({ message: 'nama tidak boleh kosong' }),
  student_father_occupation: z.string({ message: 'nama tidak boleh kosong' }),
  student_father_monthlyincome: z.string({ message: 'nama tidak boleh kosong' }),
  student_father_phonewa: z.string({ message: 'nama tidak boleh kosong' }),
  student_father_phonecallonly: z.string({ message: 'nama tidak boleh kosong' }),
  student_father_address: z.string({ message: 'nama tidak boleh kosong' }),

  student_mother_name: z.string({ message: 'nama tidak boleh kosong' }),
  student_mother_nik: z.string({ message: 'nama tidak boleh kosong' }),
  student_mother_status: z.string({ message: 'nama tidak boleh kosong' }),
  //student_mother_dob: z.string({ message: 'nama tidak boleh kosong' }),
  student_mother_deadstatus: z.string({ message: 'nama tidak boleh kosong' }),
  student_mother_lastedu: z.string({ message: 'nama tidak boleh kosong' }),
  student_mother_occupation: z.string({ message: 'nama tidak boleh kosong' }),
  student_mother_monthlyincome: z.string({ message: 'nama tidak boleh kosong' }),
  student_mother_phonewa: z.string({ message: 'nama tidak boleh kosong' }),
  student_mother_phonecallonly: z.string({ message: 'nama tidak boleh kosong' }),
  student_mother_address: z.string({ message: 'nama tidak boleh kosong' }),

  student_bear_the_costs: z.string({ message: 'nama tidak boleh kosong' }),
  student_bear_the_costs_name: z.string({ message: 'nama tidak boleh kosong' }),
  student_bear_the_costs_family: z.string({ message: 'nama tidak boleh kosong' }),
  student_bear_the_costs_phone: z.string({ message: 'nama tidak boleh kosong' }),
  student_bear_the_accupation: z.string({ message: 'nama tidak boleh kosong' }),
  student_bear_the_costs_address: z.string({ message: 'nama tidak boleh kosong' }),

  student_address: z.string({ message: 'nama tidak boleh kosong' }),
  student_rt: z.string({ message: 'nama tidak boleh kosong' }),
  student_rw: z.string({ message: 'nama tidak boleh kosong' }),
  student_province: z.string({ message: 'nama tidak boleh kosong' }),
  student_city: z.string({ message: 'nama tidak boleh kosong' }),
  student_district: z.string({ message: 'nama tidak boleh kosong' }),
  student_subdistrict: z.string({ message: 'nama tidak boleh kosong' }),
})

export type datapribadiSchema = z.infer<typeof datapribadiSchema>